Ticket_3:Code the view and script to mark readed and display 10 last requests

+++ apps/request_history/my_middleware.py
@@ -3,9 +3,12 @@ from .models import MyRequest
 
 class RequestHistoryMiddleware():
     def process_request(self, request):
-        instance = MyRequest.objects.create(request_method=request.method, request_link=request.META.get('HTTP_REFERER', ''))
-        instance = MyRequest.objects.order_by('-request_time')
-        for c in range(len(instance)):
-            if c >= 10:
-                instance[c].delete()
+        if request.is_ajax():
+            pass
+        else:
+            instance = MyRequest.objects.create(request_method=request.method, request_link=request.META.get('HTTP_REFERER', ''))
+            instance = MyRequest.objects.order_by('-request_time')
+            for c in range(len(instance)):
+                if c >= 10:
+                    instance[c].delete()
         return None
